// packages/sample-prosemirror-teemu-full/src/core/types/editor-plugin.ts
import { MarkConfig, NodeConfig } from './editor-config';
import { PMPlugin } from './pm-plugin';

export interface PluginsOptions {
  [pluginName: string]: any;
  // blockQuote?: BlockQuoteHandler;
}

/** 自定义的插件和配置单元 */
export interface EditorPlugin {
  /**
   * Name of a plugin, that other plugins can use to provide options to it or exclude via a preset.
   */
  name: string;

  /**
   * Options that will be passed to a plugin with a corresponding name if it exists and enabled.
   */
  pluginsOptions?: any;

  /**
   * List of ProseMirror-plugins. This is where we define which plugins will be added to EditorView (main-plugin, keybindings, input-rules, etc.).
   */
  pmPlugins?: (pluginOptions?: any) => PMPlugin[];

  /**
   * List of Nodes to add to the schema.
   */
  nodes?: () => NodeConfig[];

  /**
   * List of Marks to add to the schema.
   */
  marks?: () => MarkConfig[];
}